src/main/java/pro/sky/coursework2/Service/JavaQuestionService.java
package pro.sky.coursework2.Service;

import org.springframework.stereotype.Service;
import pro.sky.coursework2.domain.Question;

import java.util.*;

@Service
public class JavaQuestionService implements QuestionService {

    Set<Question> questions = new HashSet<>();

    @Override
    public Question add(String question, String answer) {
        return add(new Question(question, answer));
    }

    @Override
    public Question add(Question question) {
        questions.add(question);
        return question;
    }

    @Override
    public Question remove(Question question) {
        questions.remove(question);
        return question;
    }

    @Override
    public Collection<Question> getAll() {
        return questions;
    }

    @Override
    public Question getRandomQuestion() {
        Random random = new Random();
        List<Question> setCopy = new ArrayList<>(List.copyOf(questions));
        return setCopy.get(random.nextInt(setCopy.size()));
    }
}
